Clean up dask-remote
Move CLI logic back into distributed.cli

Refactor `submit` subpackage into `submit.py`

This includes moving the tests into the main `tests/` subdirectory
as `test_submit_remote_client` and `test_submit_cli`.

Fix test failures

Change example host in docs to localhost

Refactor signal handlers

Pull signal handling into a helper function in `distributed.cli.utils`,
and call it in the `if __name__ == '__main__'` block of the script, to
ensure there aren't side effects from importing them as modules.

Revert signal handling changes to `dask_worker`

Remove additional indirection installing signal handlers
